+++ funasr/auto/auto_tokenizer.py
@@ -1,0 +1,8 @@
+
+
+class AutoTokenizer:
+	"""
+	Undo
+	"""
+	def __init__(self):
+		pass
